Remove: 기존 component 제거
