+++ src/components/common/post/PostCard/PostCard.module.scss
