+++ src/components/common/post/PostCard/postCard.module.scss
@@ -1,11 +1,8 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@500;700&display=swap');
-$FONT: 'Noto Sans KR', sans-serif;
-
-.postCard {
+.post_card_container {
   overflow: hidden;
-  //width: 100%;
   width: 270px;
   height: 310px;
+  padding: 20px;
   transition: 0.3s;
   border-radius: 20px;
   background-color: #fff;
@@ -15,152 +12,149 @@ $FONT: 'Noto Sans KR', sans-serif;
     transform: scale(1.05);
     transition: transform 0.4s;
   }
+
+  .post_card_header {
+    min-height: 40px;
+    height: 15%;
+
+    .post_card_header_wrap {
+      display: flex;
+      overflow: hidden;
+      justify-content: space-between;
+      align-items: center;
+
+      .post_card_header_left {
+        display: flex;
+        height: 100%;
+        align-items: center;
+        gap: 8px;
+        font-weight: 700;
+        font-size: 13px;
+      }
+      .post_card_img_wrap {
+        background: #ffffff;
+        border: 1.5px solid rgba(0, 0, 0, 0.5);
+        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
+        border-radius: 50%;
+        overflow: hidden;
+      }
+    }
+  }
+
+  .post_card_center {
+    height: 68%;
+    padding: 20px 7px;
+    overflow: hidden;
+    .post_card_title {
+      width: 100%;
+      font-weight: 500;
+      font-size: 18px;
+      margin-bottom: 7px;
+      & * {
+        width: 100%;
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+      }
+    }
+    .post_card_content {
+      height: 60%;
+      width: 100%;
+      overflow: hidden;
+      & * {
+        width: 100%;
+        height: 60%;
+        word-break: break-word;
+        font-size: 0.875rem;
+        line-height: 1.4;
+        display: -webkit-box;
+        -webkit-line-clamp: 7;
+        -webkit-box-orient: vertical;
+        overflow: hidden;
+        text-overflow: ellipsis;
+      }
+    }
+    .post_card_skills_wrap {
+      height: 40%;
+      .post_card_skills {
+        display: flex;
+        gap: 10px;
+        justify-content: center;
+
+        .post_card_skill {
+          background: #ffffff;
+          border-radius: 50%;
+          box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.15);
+        }
+
+        .post_card_skill_over_wrap {
+          position: relative;
+          width: 20px;
+          height: 20px;
+          .post_card_skill_over {
+            position: absolute;
+            top: 0;
+            left: -100%;
+            background: #ffffff;
+            box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.15);
+            border-radius: 50%;
+          }
+        }
+        .post_card_skills_ellipsis {
+          & img {
+            height: 20px;
+            width: 14px;
+            margin-top: 10px;
+          }
+        }
+      }
+    }
+  }
+
+  .post_card_footer {
+    height: 17%;
+    padding: 0 10%;
+    overflow: hidden;
+
+    .post_card_footer_items {
+      height: 100%;
+      display: flex;
+      justify-content: space-around;
+      align-items: flex-end;
+      gap: 15px;
+
+      .post_card_footer_item {
+        display: flex;
+        align-items: center;
+        gap: 6px;
+        color: #848484;
+        font-weight: 400;
+        font-size: 14px;
+      }
+    }
+  }
 }
 
-.postWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
-}
-
-.postImgWrapper {
-  height: 50px;
-  //background-color: #568a35;
-}
-
-.postImgBox {
-  position: relative;
-  display: flex;
-  top: 45%;
-  left: 10%;
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  box-shadow: 0 0 2px rgba(0, 0, 0, 1);
-  background-color: #fff;
-  //border: 1px solid rgba(0, 0, 0, 0.5);
-}
-
-.img {
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-}
-
-.postContentBox {
-  padding-top: 45px;
-  height: 100%;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  justify-content: space-between;
-}
-
-.postStatusWrapper {
-  height: fit-content;
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-}
-
-.postStatus {
+.status_badge {
   width: 60px;
   height: 25px;
-  margin-right: 25px;
-  margin-top: -11px;
-  padding: 5px 12px;
-  border-radius: 20px;
-  background-color: #6ccb6b;
-  //background-color: &{({ isDone }) => (isDone ? '#878787' : '#0ac858')};
+  text-align: center;
+  line-height: 22px;
   color: #fff;
-  font-family: $FONT;
-  font-weight: bold;
+  font-weight: 700;
   font-size: 13px;
+
+  background: #6ccb6b;
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
+  border-radius: 20px;
 }
 
-.postStatusNo {
-  width: 65px;
-  height: 25px;
-  margin-right: 25px;
-  margin-top: -11px;
-  padding: 5px 8.5px;
-  border-radius: 20px;
+.status_badge_no {
   background-color: #878787;
   color: #fff;
-  font-family: $FONT;
-  font-weight: bold;
-  font-size: 13px;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
 }
-
-.postTitle {
-  padding: 0 35px;
-  font-size: 18px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  font-family: $FONT;
-  font-weight: normal;
-}
-
-.postContent {
-  padding: 0 35px;
-  margin-top: 20px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  line-height: 150%;
-  font-family: $FONT;
-  font-weight: normal;
-  font-size: 15px;
-  color: #3e3e3e;
-}
-
-.postInfoBox {
-  width: 170px;
-  height: 40px;
-  margin-bottom: 10px;
-}
-
-.postInfoLine {
+.post_info_line {
   width: 200px;
   border-top: 1px solid #e9e9e9;
-  margin: 0 0 10px 35px;
-}
-
-.postInfo {
-  font-family: $FONT;
-  display: flex;
-  align-items: center;
-  font-size: 13px;
-  font-weight: bold;
-  margin-left: 20px;
-  gap: 4px;
-  color: #848484;
-}
-
-.postInfoName {
-  font-family: $FONT;
-  display: flex;
-  align-items: center;
-  font-size: 14px;
-  color: #393939;
-  font-weight: 700;
-  margin-left: 12px;
-}
-
-.postIconBox {
-  display: flex;
-  //justify-content: flex-end;
-  margin-left: 44px;
-  > img {
-    margin-top: 1px;
-  }
+  margin: 0 auto;
 }
