--- src/components/pages/CreateStudyPage/CreateStudyPage.tsx
@@ -1,8 +1,0 @@
-import { CreateStudy } from '@components/organism/CreateStudy/CreateStudy';
-// 클릭한 값 보여주기
-
-function CreateStudyPage() {
-  return <CreateStudy />;
-}
-
-export default CreateStudyPage;
