--- src/components/pages/DetailPage/DetailPage.scss
@@ -1,277 +1,0 @@
-.detail_container {
-  width: 78%;
-  margin: 50px auto;
-  padding-left: 13%;
-  padding-right: 13%;
-}
-
-.detail_profile_wrapper {
-  display: flex;
-}
-
-.detail_arrow_wrapper {
-  margin-right: 50px;
-  margin-top: 25px;
-}
-
-.detail_arrow {
-  // position: absolute;
-  width: 40px;
-  height: 40px;
-  text-align: center;
-  // left: 80px;
-  cursor: pointer;
-}
-
-.detail_profile {
-  width: 80px;
-  height: 80px;
-  border: 3px solid #98b487;
-  // margin-bottom: 30px;
-  // margin-left: 5px;
-  background-size: contain;
-  background-repeat: no-repeat;
-  background-position: center;
-  border-radius: 50%;
-}
-.detail_setting{
-  position: relative;
-}
-.detail-toggle-container{
-  position:absolute;
-  top:0;
-  left: 40px;
-  width:150px;
-  height:90px;
-  background: #FFFFFF;
-  border: 1px solid #BFBFBF;
-  border-radius: 10px;
-  box-shadow: 0px 4px 4px rgb(0 0 0 / 10%);
-  border-radius: 5px;
-  z-index: 2;
-  color: rgba(140, 140, 140, 1);
-
-
-  li{
-    width:100%;
-    height:45px;
-    line-height:45px;
-    text-align: center;
-    &:hover{
-      cursor: pointer;
-      box-shadow: inset 0 1000px rgba(0, 0, 0, 0.03) !important;
-
-    }
-    &:last-child{
-      border-top: 1px solid #E6E6E6;
-    }
-    >span{
-      color: rgba(83, 83, 83, 1); 
-    }
-  }
-}
-
-.detail_summary_container {
-  width: 80%;
-  padding-left: 20px;
-}
-
-.summary_title_line {
-  border: 1px solid #568a35;
-  width: auto;
-  margin: auto auto;
-}
-
-.summary_title_text {
-  font-weight: 500;
-  font-size: 30px;
-  line-height: 43px;
-  color: #333333;
-  margin: 10px 0px 15px 25px;
-  padding-left: 15px;
-  padding-top: 10px;
-}
-
-.summary_title_addoption {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.summary_title_right{
-  display: flex;
-  gap: 25px;
-}
-
-.summary_title_icon {
-  width: 30px;
-  height: 30px;
-  margin-top: 8px;
-  margin-right: 20px;
-  &:hover{
-    cursor: pointer;
-  }
-}
-
-.summary_category_wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.summary_category_left {
-  display: flex;
-  padding-left: 10px;
-}
-
-@mixin category($bgcolor, $bordercolor, $fontcolor) {
-  width: 100px;
-  height: 40px;
-  border-radius: 8px;
-  margin: 15px 2px 60px 13px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1.5px solid $bordercolor;
-  background-color: $bgcolor;
-  color: $fontcolor;
-}
-
-@mixin hover(){
-  &:hover {
-    cursor: pointer;
-    transform: scale(1.2);
-    transition: transform 0.5s;
-  }
-}
-.summary_category_name {
-  @include category(white, #568a35, #568a35);
-}
-
-.summary_category_status_true {
-  @include category(#f09933, #f09933, white);
-  &.hover{
-   @include hover();
-  }
-}
-
-.summary_category_status_false {
-  @include category(gray, gray, white);
-  &.hover{
-    @include hover();
-   }
-}
-
-.summary_category_right {
-  display: flex;
-  align-items: center;
-}
-
-.summary_like_true_button {
-  width: 30px;
-  height: 26px;
-  margin-top: 10px;
-  cursor: pointer;
-  &:hover {
-    transform: scale(1.2);
-    transition: transform 0.5s;
-  }
-}
-
-.summary_like_false_button {
-  width: 30px;
-  height: 26px;
-  margin-top: 10px;
-  filter: opacity(0.5) drop-shadow(0 0 0 gray) saturate(0%);
-  cursor: pointer;
-  &:hover {
-    transform: scale(1.2);
-    transition: transform 0.4s;
-  }
-}
-
-.detail_sort_wrapper {
-  display: flex;
-  margin-left: 215px;
-  font-size: 18px;
-  font-weight: 300;
-}
-
-.detail_sort_position{
-  display: inline-block;
-  margin-right: 10px;
-  font-size: 16px;
-}
-
-.detail_sort_content {
-  margin: 30px 15px 16px -100px;
-  color: #555555;
-}
-
-.detail_contact {
-  color: #555555;
-  margin-left: 115px;
-  font-size: 18px;
-  font-weight: 300;
-  a{
-    color: #6a77ed;
-    font-weight: 400;
-  }
-}
-
-.detail_line {
-  width: 100%;
-  border: 1.5px solid #c2c2c2;
-  margin: 90px 0px;
-}
-
-.detail_content_wrapper {
-  margin: 30px;
-  height: fit-content;
-}
-
-.detail_content_title {
-  margin: 20px 40px 30px 40px;
-  font-weight: 500;
-  font-size: 24px;
-  line-height: 30px;
-}
-
-.detail_content {
-  color: #8c8c8c;
-  padding: 0% 3% 0% 3%;
-  margin: 40px;
-  line-height: 23px;
-  height: max-content;
-  min-height: 500px;
-}
-
-.detail_reaction {
-  display: flex;
-  justify-content: end;
-}
-
-.reaction_content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 40px;
-  margin: 5px 10px;
-}
-
-.reaction_image {
-  width: 20px;
-  margin: 0px 10px;
-}
-.comment_title {
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 29px;
-  color: #6a6a6a;
-  width: 100%;
-  text-align: left;
-  margin: 6px 0px 35px 0px;
-}
-.comment_list {
-  display: grid;
-  row-gap: 28px;
-}
