--- src/components/pages/MainPage/MainPage.scss
@@ -1,14 +1,0 @@
-.mainPage_section_wrapper {
-  .study_section {
-    width: 1420px;
-    margin: 0 auto;
-    padding-bottom: 70px;
-  }
-  .body_section {
-    width: 1420px;
-    margin: 0 auto;
-  }
-  .search {
-    padding: 30px 0 65px 0;
-  }
-}
