--- src/components/pages/MyPage/MyPage.scss
@@ -1,127 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@300;400;500;700&display=swap');
-
-.myPageWrapper{
-  text-align: center;
-  margin: 0 auto 100px auto;
-}
-
-
-.myPageBackground{
-  margin-top: -30px;
-  height: 250px;
-  width: 100%;
-  background-color: #D9D9D9;
-  //z-index: -1;
-  position: relative;
-}
-
-
-.myPageProfileWrapper{
-
-}
-.myProfileImg{
-  border-radius: 50%;
-  width: 170px;
-  height: 170px;
-  margin-top: 130px;
-}
-.myNickName{
-  margin: 70px auto 12px auto;
-  font-weight: 700;
-  font-size: 36px;
-  color: #484848;
-}
-.myInfo{
-  margin: 5px;
-  font-weight: 400;
-  font-size: 16px;
-  color: #767676;
-}
-.myIntro{
-  width: 42%;
-  height: 30px;
-  text-align: center;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 20px;
-  color: #767676;
-  margin: 25px auto;
-}
-
-
-.myPositionWrapper{
-  position: relative;
-  height: 50px;
-  width: 100%;
-}
-.myPosition{
-  top: 50%;
-  margin: 50px auto auto auto;
-  font-weight: 700;
-  font-size: 15px;
-  width: fit-content;
-  padding: 10px;
-  height: 35px;
-  color: #FFFFFF;
-  background: #000000;
-  border-radius: 5px;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-
-
-.myPostsWrapper{
-  position: relative;
-  height: 50px;
-  font-weight: 500;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  color: #FFFFFF;
-}
-.myPosts{
-  width: 210px;
-  height: 50px;
-  background: #FFFFFF;
-  border: 1.5px solid #898989;
-  border-radius: 10px;
-  color: #898989;
-  padding: 15px;
-  //margin-right: 25px;
-  cursor: pointer;
-  margin: 40px 10px auto auto;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-.myStudy{
-  width: 210px;
-  height: 50px;
-  background: #568A35;
-  border-radius: 10px;
-  padding: 15px;
-  cursor: pointer;
-  margin: 40px auto auto 10px;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-
-
-.myLine{
-  margin-top: 130px;
-  width: 100%;
-  height: 0;
-  border: 1px solid #D9D9D9;
-}
-
-
-.myEditProfile{
-  padding: 6px 20px;
-  margin: -25px auto auto 73%;
-  border-radius: 30px;
-  border: 0.7px solid #898989;
-  font-family: 'Noto Sans KR', sans-serif;
-  font-weight: 700;
-  font-size: 15px;
-  display: flex;
-  align-items: center;
-  color: #898989;
-  background-color: #fff;
-  cursor: pointer;
-}
