--- src/components/pages/NotMatchPage/style.module.scss
@@ -1,18 +1,0 @@
-.NotMatchContainer {
-  width: 100vw;
-  height: 50vh;
-
-  padding: 25% 20%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  gap: 20px;
-
-  .Title {
-    font-size: 20px;
-  }
-  .NavButton {
-  }
-}
