--- src/components/pages/index.tsx
@@ -1,2 +1,0 @@
-export * from './MainPage/MainPage';
-export * from './PostsPage';
